requirements.txt
Flask==1.1.1
certifi==2020.4.5.1
click==7.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
MarkupSafe==1.1.1
pip==20.0.2
pycrypto==2.6.1
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
Werkzeug==1.0.0
wheel==0.34.2
wincertstore==0.2
gunicorn==19.5.0
